feature: writing basic booking management logic to room and inn models https://github.com/anEffingChamp/oviaHealth
Guests can not store their luggage in another guest’s room.

Being that this is a shared space inn, guests might be in shared rooms if that’s
the most profitable solution.

The cost is calculated per person according to this formula:
(Base room cost / number of people in the room ) + (base storage costs * number
of items stored).
Base room cost is 10 Gold, storage cost is 2 Gold.

reference:
https://docs.google.com/document/d/1f6Atif7Y08C4AjAbTlqk-_3O1SYy55swKj4BN2Vk55I/edit?ts=5bedd786